Add files via upload
A non-negative "large integer" (i.e. - integers outside of Java's int range) is represented as an array of N digits, where the value of each digit is an integer from 0 to 9. Zero values at the beginning of the array indicate leading zeroes. 

This mini-program takes two user inputs being "large integers" and uses arrays to calculate the sum.